pyproject.toml
[project]
name = "riffs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-taggit>=6.1.0",
    "whitenoise>=6.9.0",
]
